updated to allow for map selection
added the ability to select a station by clicking on the marker on the map with the observeEvent() function. Clicking on a table or map marker replaces the previous selection. Temporarily removed the marker popup on click (disappears with map reloading). Had to put the Station Listings on the first tab as it needs to render first (if a map selection is made before the listings table has had a chance to render, the selected station goes back to default. doesnt occur after rendering the first time).